+++ common/replay/replayutils.h
@@ -10,8 +10,8 @@
 
 #include "utlstring.h"
 
-void Replay_GetFirstAvailableFilename( char *pDst, int nDstLen, const char *pIdealFilename, const char *pExt,
-									   const char *pFilePath, int nStartIndex );
+void Replay_GetFirstAvailableFilename( char *pDst, intp nDstLen, const char *pIdealFilename, const char *pExt,
+									   const char *pFilePath, intp nStartIndex );
 
 void Replay_ConstructReplayFilenameString( CUtlString &strOut, const char *pReplaySubDir, const char *pFilename, const char *pGameDir );
